requirements.txt
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.1.0
fastapi==0.119.0
uvicorn==0.37.0
kafka-python==2.2.15
